ICPP 2015 version ready

+++ bin/dewe.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-
+hostname=$(hostname)
+datetime=$(date +"%Y-%m-%d-%H-%M-%S")
 #
 # This part starts / stops the master daemon
 #
@@ -10,13 +11,15 @@
 if [ $1 == "master" ]; then
     if [ $2 == "start" ]; then
         echo "start master"
-        nohup java -jar ../target/DEWE.v2.one-jar.jar master < /dev/null > master.log 2>&1 &
+        nohup java -jar ../target/DEWE.v2.one-jar.jar master < /dev/null > master-$hostname-$datetime.log 2>&1 &
         echo $! > master.pid
     fi
     if [ $2 == "stop" ]; then
         echo "stop master"
-        kill -TERM $(cat master.pid)
-        rm master.pid
+        if [ -f 'master.pid' ]; then
+          kill -TERM $(cat master.pid)
+          rm master.pid
+        fi
     fi
 fi
 
@@ -30,13 +33,15 @@ fi
 if [ $1 == "worker" ]; then
     if [ $2 == "start" ]; then
         echo "start worker"
-        nohup java -jar ../target/DEWE.v2.one-jar.jar worker < /dev/null > worker.log 2>&1 &
+        nohup java -jar ../target/DEWE.v2.one-jar.jar worker < /dev/null > worker-$hostname-$datetime.log 2>&1 &
         echo $! > worker.pid
     fi
     if [ $2 == "stop" ]; then
         echo "stop worker"
-        kill -TERM $(cat worker.pid)
-        rm worker.pid
+        if [ -f 'worker.pid' ]; then
+          kill -TERM $(cat worker.pid)
+          rm worker.pid
+        fi
     fi
 fi
 
@@ -51,13 +56,15 @@ if [ $1 == "monitor" ]; then
     if [ -z "$3" ]; then
       if [ $2 == "stop" ]; then
           echo "stop load monitor"
-          kill -TERM $(cat monitor.pid)
-          rm monitor.pid
+          if [ -f 'monitor.pid' ]; then
+            kill -TERM $(cat monitor.pid)
+            rm monitor.pid
+          fi
       fi
     else
       if [ $3 == "start" ]; then
           echo "start load monitor"
-          nohup java -jar ../target/DEWE.v2.one-jar.jar monitor $2 < /dev/null > monitor.log 2>&1 &
+          nohup java -jar ../target/DEWE.v2.one-jar.jar monitor $2 < /dev/null > monitor-$hostname-$datetime.log 2>&1 &
           echo $! > monitor.pid
       fi
     fi
